Route all terminal IO through a Log class
`Kernel.gets` & `puts` is fine, but it's not very testable. What we 
really want is to capture all input/output in each test, so we'll set up
a fake IO object for each test

This isn't the end-all design (because it's big and global), but
I wanted to make the minimal seam possible to get the first test going
and didn't want to introduce a complex dependency-injection scheme or
start pulling apart concerns before the test is in place.