Found an errant reference to home which I renamed
As a general rule, stuff in a script/ directory is presumed to depend
on stuff in lib/ (and it clearly does), but when stuff in lib also
depends back on script/,  what happens is called a circular dependency
or in some parlance, a dependency tangle.

This is considered bad form because dependencies are much easier to 
think about when they're unidirectional. "The driver depends on the 
browser depends on the toolbox" makes sense whereas "the driver depends
on the browser depends on the toolbox which also depends on the browser"
is much harder to deal with. For starters:

* The Toolbox can't be tested without the Driver
* If both the Toolbox and the Driver do stuff at load-time, the order
they are required will matter and it's possible you'll get stuck in an
infinite loop or an impossible situation from a load ordering perspective